Added spec
